refactor(image-web): renaming rendering component
