+++ packages/pluggableWidgets/image-web/src/components/Image/Image.tsx
